Commit 13 - 2023/04/28

+++ apis/delete-user-api.php
@@ -7,35 +7,50 @@ if (isset($_SESSION["user-is-admin"])) {
     $userId = filter_var($_POST["user-id"], FILTER_SANITIZE_FULL_SPECIAL_CHARS);
 
     if ($userId) {
-        $chatsPath =   "../admin/chat-channels/";
 
-        $chats = array();
-        if (is_dir($chatsPath)) {
-            $chatsArray = array_diff(scandir($chatsPath), array('.', '..'));
-            /* Create a new array to reorganize indexes */
 
-            $chats = array_values($chatsArray);
+        $mySqlUserQuery = "SELECT * FROM users WHERE user_id = $userId";
+        $mySqlUserResult = mysqli_query($connection, $mySqlUserQuery);
+        if (mysqli_num_rows($mySqlUserResult) == 1) {
+            $mySqlUser = mysqli_fetch_assoc($mySqlUserResult);
+            unlink("../admin/received-files/avatars/" . $mySqlUser["avatar"]);
 
-            sort($chats);
-            $i = 0;
-            foreach ($chats as $index => $chat) {
-                $chatExtension = explode(".", $chat);
-                if (end($chatExtension) == "json") {
-                    //Read The Chat File
-                    $chatObj = json_decode(file_get_contents($chatsPath . $chat), false);
+            $chatsPath =   "../admin/chat-channels/";
 
-                    foreach ($chatObj->messages as $index2 => $messages) {
-                        if ($messages->userId == $userId) {
-                            unset($chatObj->messages[$index2]);
+            $chats = array();
+            if (is_dir($chatsPath)) {
+                $chatsArray = array_diff(scandir($chatsPath), array('.', '..'));
+                /* Create a new array to reorganize indexes */
+
+                $chats = array_values($chatsArray);
+
+                sort($chats);
+                $i = 0;
+                foreach ($chats as $index => $chat) {
+                    $chatExtension = explode(".", $chat);
+                    if (end($chatExtension) == "json") {
+                        //Read The Chat File
+                        $chatObj = json_decode(file_get_contents($chatsPath . $chat), false);
+
+                        foreach ($chatObj->messages as $index2 => $messages) {
+                            if ($messages->userId == $userId) {
+                                if ($messages->messageMedia != "") {
+                                    unlink("../admin/received-files/chat-media/" . $messages->messageMedia);
+                                }
+                                unset($chatObj->messages[$index2]);
+                            }
                         }
+
+
+                        file_put_contents($chatsPath . $chat, json_encode($chatObj));
                     }
-
-
-                    file_put_contents($chatsPath . $chat, json_encode($chatObj));
                 }
+
+                $mysqlQuery = "DELETE FROM users WHERE user_id = $userId";
+                $mysqlResult = mysqli_query($connection, $mysqlQuery);
+            } else {
+                $deleteUserStatus = 4;
             }
-            $mysqlQuery = "DELETE FROM users WHERE user_id = $userId";
-            $mysqlResult = mysqli_query($connection, $mysqlQuery);
         } else {
             $deleteUserStatus = 3;
         }
